+++ AltPro.BackTracker/ViewModels/HomeReportViewModel.cs
@@ -1,0 +1,14 @@
+﻿using AltPro.BackTracker.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AltPro.BackTracker.ViewModels
+{
+    public class HomeReportViewModel
+    {
+        public Report Report { get; set; }
+        public string PageTitle { get; set; }
+    }
+}
