--- AltPro.BackTracker/Views/Home/Privacy.cshtml
@@ -1,6 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Privacy Policy";
-}
-<h1>@ViewData["Title"]</h1>
-
-<p>Use this page to detail your site's privacy policy.</p>
