+++ AltPro.BackTracker/Views/Home/ReportList.cshtml
@@ -6,3 +6,29 @@
 
 <form asp-controller="Home" asp-action="ListReport" method="post"></form>
 
+<div>
+    @foreach (var report in Model)
+    {
+    <div class="card m-1">
+        <a href="#" class="list-group-item list-group-item-action flex-column align-items-start">
+            <div class="d-flex w-100 justify-content-between">
+                <h5 class="mb-1">@report.Title</h5>
+                <small class="text-muted">@report.Time</small>
+            </div>
+            <div>
+                <small class="text-muted"><strong>Priority: @report.Priority</strong>,</small>
+                <small class="text-muted"><strong>Status: @report.Status</strong></small>
+            </div>
+
+            <small class="text-muted">Assigness: @report.Assigness</small>
+            <p>
+                <div>
+                    <small herf="#" class="btn btn-primary m-1">View</small>
+                    <small herf="#" class="btn btn-primary m-1">Edit</small>
+                    <small herf="#" class="btn btn-danger m-1">Delete</small>
+                </div>
+            </p>
+        </a>
+    </div>
+    }
+</div>
